fix new labels layer creation

+++ napari/components/_viewer/model.py
@@ -174,9 +174,9 @@ class Viewer:
 
     def _new_labels(self):
         if self.dims.max_dims == 0:
-            empty_labels = np.zeros((512, 512))
+            empty_labels = np.zeros((512, 512), dtype=int)
         else:
-            empty_labels = np.zeros(self.dims.max_shape)
+            empty_labels = np.zeros(self.dims.max_shape, dtype=int)
         self.add_labels(empty_labels)
 
     def _update_layers(self):
